reduce utils, comment about GPTQ
