+++ pyproject.toml
@@ -197,4 +197,4 @@ classmethod-decorators = ["classmethod", "olive.common.pydantic_v1.validator", "
 "scripts/**" = ["INP001"]
 "examples/directml/llm/chat_app/**" = ["TID252", "UP006", "T201"]
 "olive/cli/**" = ["T201"]
-"olive/**/hadamard_tensors.py" = ["E501"]
+"olive/**/hadamard_utils.py" = ["E501"]
