+++ test/unit_test/passes/pytorch/test_torch_trt_conversion.py
@@ -13,8 +13,8 @@ from olive.common.utils import get_attr
 from olive.data.template import huggingface_data_config_template
 from olive.model import HfModelHandler
 from olive.passes.olive_pass import create_pass_from_dict
+from olive.passes.pytorch.sparsegpt_utils import get_layer_submodules
 from olive.passes.pytorch.torch_trt_conversion import TorchTRTConversion
-from olive.passes.pytorch.utils.sparsegpt import get_layer_submodules
 
 # pylint: disable=abstract-method
